requirements.txt
# requirements.txt - Dubai Cons AI Suite MVP (Python 3.10)

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML Libraries (ProxyAPI compatible)
google-generativeai>=0.3.0
openai==1.3.0
httpx==0.25.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Image Processing
pillow==10.0.1
opencv-python==4.8.1.78
imageio==2.31.6

# 3D/VR Libraries
trimesh==4.0.5
open3d==0.18.0
pywavefront==1.3.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Background Tasks
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
reportlab==4.1.0